+++ headers/text.h
@@ -4,11 +4,11 @@
 #include "texture.h"
 #include "colors.h"
 
-const float SCALE_FACTOR = 1;
-const float BIG_FONT_SCALE = 0.7;
-const float NORMAL_FONT_SCALE = 0.6;
-const float SMALL_FONT_SCALE = 0.3;
-const float FONT_SCALE = 0.5;
+const double SCALE_FACTOR = 1;
+const double BIG_FONT_SCALE = 0.7;
+const double NORMAL_FONT_SCALE = 0.6;
+const double SMALL_FONT_SCALE = 0.3;
+const double FONT_SCALE = 0.5;
 
 const int FONT_WIDTH = 16;
 const int FONT_HEIGHT = 8;
@@ -21,7 +21,7 @@ public:
 
   bool buildFont(std::string path, SDL_Window *window, SDL_Renderer *renderer);
   void renderText(int x, int y, std::string_view text, SDL_Renderer *renderer,
-                  float scale_factor = SCALE_FACTOR, int width = 0,
+                  double scale_factor = SCALE_FACTOR, int width = 0,
                   int height = 0);
   int getWordWidth(std::string_view word);
   int getWordHeight(std::string_view word);
