+++ src/edge.cpp
@@ -6,12 +6,12 @@ void Edge::draw(SDL_Renderer *renderer) const {
                             m_endNode->getX() - m_startNode->getX());
   double endAngle = startAngle + M_PI;
 
-  int startX = m_startNode->getX() + NODE_RADIUS * cos(startAngle);
-  int startY = m_startNode->getY() + NODE_RADIUS * sin(startAngle);
-  int endX = m_endNode->getX() + NODE_RADIUS * cos(endAngle);
-  int endY = m_endNode->getY() + NODE_RADIUS * sin(endAngle);
+  double startX = m_startNode->getX() + NODE_RADIUS * cos(startAngle);
+  double startY = m_startNode->getY() + NODE_RADIUS * sin(startAngle);
+  double endX = m_endNode->getX() + NODE_RADIUS * cos(endAngle);
+  double endY = m_endNode->getY() + NODE_RADIUS * sin(endAngle);
 
-  SDL_RenderDrawLine(renderer, startX, startY, endX, endY);
+  SDL_RenderDrawLine(renderer, (int)startX, (int)startY, (int)endX, (int)endY);
   if (m_oriented) {
     // calculate arrowhead points
     double angleDiff = M_PI / 6;
@@ -20,8 +20,8 @@ void Edge::draw(SDL_Renderer *renderer) const {
     double x2 = endX - NODE_RADIUS * cos(startAngle + angleDiff);
     double y2 = endY - NODE_RADIUS * sin(startAngle + angleDiff);
 
-    SDL_RenderDrawLine(renderer, endX, endY, (int)x1, (int)y1);
-    SDL_RenderDrawLine(renderer, endX, endY, (int)x2, (int)y2);
+    SDL_RenderDrawLine(renderer, (int)endX, (int)endY, (int)x1, (int)y1);
+    SDL_RenderDrawLine(renderer, (int)endX, (int)endY, (int)x2, (int)y2);
   }
 }
